services/transcription-worker/requirements.txt
redis
google-cloud-speech
google-cloud-storage
